+++ operator-common/src/main/java/io/strimzi/operator/common/operator/resource/AbstractNamespacedResourceOperator.java
@@ -9,6 +9,7 @@ import io.fabric8.kubernetes.api.model.HasMetadata;
 import io.fabric8.kubernetes.api.model.KubernetesResourceList;
 import io.fabric8.kubernetes.api.model.LabelSelector;
 import io.fabric8.kubernetes.client.KubernetesClient;
+import io.fabric8.kubernetes.client.KubernetesClientException;
 import io.fabric8.kubernetes.client.Watcher;
 import io.fabric8.kubernetes.client.dsl.FilterWatchListDeletable;
 import io.fabric8.kubernetes.client.dsl.Informable;
@@ -53,14 +54,28 @@ public abstract class AbstractNamespacedResourceOperator<C extends KubernetesCli
      */
     public final static String ANY_NAMESPACE = "*";
 
+    private final boolean useServerSideApply;
+
     /**
      * Constructor.
      * @param vertx The vertx instance.
      * @param client The kubernetes client.
      * @param resourceKind The mind of Kubernetes resource (used for logging).
+     * @param useServerSideApply whether to use server side apply patch requests
+     */
+    public AbstractNamespacedResourceOperator(Vertx vertx, C client, String resourceKind, boolean useServerSideApply) {
+        super(vertx, client, resourceKind);
+        this.useServerSideApply = useServerSideApply;
+    }
+
+    /**
+     * Constructor.
+     * @param vertx The vertx instance.
+     * @param client The kubernetes client.
+     * @param resourceKind The kind of Kubernetes resource (used for logging).
      */
     public AbstractNamespacedResourceOperator(Vertx vertx, C client, String resourceKind) {
-        super(vertx, client, resourceKind);
+        this(vertx, client, resourceKind, false);
     }
 
     protected abstract MixedOperation<T, L, R> operation();
@@ -97,27 +112,42 @@ public abstract class AbstractNamespacedResourceOperator<C extends KubernetesCli
             return Future.failedFuture("Given name " + name + " incompatible with desired name " + desired.getMetadata().getName());
         }
 
-        return getAsync(namespace, name)
-                .compose(current -> {
-                    if (desired != null) {
-                        if (current == null) {
-                            LOGGER.debugCr(reconciliation, "{} {}/{} does not exist, creating it", resourceKind, namespace, name);
-                            return internalCreate(reconciliation, namespace, name, desired);
+        if (useServerSideApply) {
+            if (desired != null) {
+                if (desired.getMetadata().getManagedFields() != null && !desired.getMetadata().getManagedFields().isEmpty()) {
+                    LOGGER.debugCr(reconciliation, "Deleting managedFields from request before pathing resource {} {}/{}", resourceKind, namespace, name);
+                    desired.getMetadata().setManagedFields(null);
+                }
+
+                LOGGER.debugCr(reconciliation, "{} {}/{} desired, patching it", resourceKind, namespace, name);
+                return internalPatch(reconciliation, namespace, name, desired);
+            } else {
+                LOGGER.debugCr(reconciliation, "{} {}/{} no longer desired, deleting it", resourceKind, namespace, name);
+                return internalDelete(reconciliation, namespace, name);
+            }
+        } else {
+            return getAsync(namespace, name)
+                    .compose(current -> {
+                        if (desired != null) {
+                            if (current == null) {
+                                LOGGER.debugCr(reconciliation, "{} {}/{} does not exist, creating it", resourceKind, namespace, name);
+                                return internalCreate(reconciliation, namespace, name, desired);
+                            } else {
+                                LOGGER.debugCr(reconciliation, "{} {}/{} already exists, updating it", resourceKind, namespace, name);
+                                return internalUpdate(reconciliation, namespace, name, current, desired);
+                            }
                         } else {
-                            LOGGER.debugCr(reconciliation, "{} {}/{} already exists, updating it", resourceKind, namespace, name);
-                            return internalUpdate(reconciliation, namespace, name, current, desired);
+                            if (current != null) {
+                                // Deletion is desired
+                                LOGGER.debugCr(reconciliation, "{} {}/{} exist, deleting it", resourceKind, namespace, name);
+                                return internalDelete(reconciliation, namespace, name);
+                            } else {
+                                LOGGER.debugCr(reconciliation, "{} {}/{} does not exist, noop", resourceKind, namespace, name);
+                                return Future.succeededFuture(ReconcileResult.noop(null));
+                            }
                         }
-                    } else {
-                        if (current != null) {
-                            // Deletion is desired
-                            LOGGER.debugCr(reconciliation, "{} {}/{} exist, deleting it", resourceKind, namespace, name);
-                            return internalDelete(reconciliation, namespace, name);
-                        } else {
-                            LOGGER.debugCr(reconciliation, "{} {}/{} does not exist, noop", resourceKind, namespace, name);
-                            return Future.succeededFuture(ReconcileResult.noop(null));
-                        }
-                    }
-                });
+                    });
+        }
     }
 
     /**
@@ -248,6 +278,17 @@ public abstract class AbstractNamespacedResourceOperator<C extends KubernetesCli
         }
     }
 
+    protected Future<ReconcileResult<T>> internalPatch(Reconciliation reconciliation, String namespace, String name, T desired) {
+        try {
+            T result = patch(reconciliation, namespace, name, desired);
+            LOGGER.debugCr(reconciliation, "{} {} in namespace {} has been patched", resourceKind, name, namespace);
+            return Future.succeededFuture(ReconcileResult.patchedUsingServerSideApply(result));
+        } catch (Exception e) {
+            LOGGER.debugCr(reconciliation, "Caught exception while patching {} {} in namespace {}", resourceKind, name, namespace, e);
+            return Future.failedFuture(e);
+        }
+    }
+
     /**
      * Method for patching or replacing a resource. By default, is using JSON-type patch. Overriding this method can be
      * used to use replace instead of patch or different patch strategies.
@@ -260,6 +301,23 @@ public abstract class AbstractNamespacedResourceOperator<C extends KubernetesCli
      */
     protected T patchOrReplace(String namespace, String name, T desired)   {
         return operation().inNamespace(namespace).withName(name).patch(PatchContext.of(PatchType.JSON), desired);
+    }
+
+    protected T patch(Reconciliation reconciliation, String namespace, String name, T desired) {
+        try {
+            return operation().inNamespace(namespace).withName(name).patch(serverSideApplyPatchContext(false), desired);
+        } catch (KubernetesClientException e) {
+            LOGGER.warnCr(reconciliation, "{} {} in namespace {} failed to apply, using force", resourceKind, name, namespace, e);
+            return operation().inNamespace(namespace).withName(name).patch(serverSideApplyPatchContext(true), desired);
+        }
+    }
+
+    protected PatchContext serverSideApplyPatchContext(boolean force) {
+        return new PatchContext.Builder()
+                .withPatchType(PatchType.SERVER_SIDE_APPLY)
+                .withFieldManager("strimzi-cluster-operator")
+                .withForce(force)
+                .build();
     }
 
     /**