+++ operator-common/src/main/java/io/strimzi/operator/common/operator/resource/PodOperator.java
@@ -28,9 +28,19 @@ public class PodOperator extends AbstractReadyNamespacedResourceOperator<Kuberne
      * Constructor
      * @param vertx The Vertx instance
      * @param client The Kubernetes client
+     * @param useServerSideApply Whether to use server side apply
+     */
+    public PodOperator(Vertx vertx, KubernetesClient client, boolean useServerSideApply) {
+        super(vertx, client, "Pods", useServerSideApply);
+    }
+
+    /**
+     * Constructor
+     * @param vertx The Vertx instance
+     * @param client The Kubernetes client
      */
     public PodOperator(Vertx vertx, KubernetesClient client) {
-        super(vertx, client, "Pods");
+        super(vertx, client, "Pods", false);
     }
 
     @Override